Moved CloseBuilderDialog into controllers
